Reshuffled stuff around + simplified dom-ancestor-visitor.
