+++ ads/google/a4a/test/test-utils.js
@@ -29,7 +29,6 @@ import {
   getCorrelator,
   getCsiAmpAnalyticsVariables,
   getEnclosingContainerTypes,
-  getFlexibleAdSlotRequestParams,
   getIdentityToken,
   getIdentityTokenRequestUrl,
   googleAdUrl,
@@ -1121,143 +1120,6 @@ describes.realWin('#groupAmpAdsByType', {amp: true}, env => {
             )
           )
       );
-    });
-  });
-});
-
-describes.realWin('#getFlexibleAdSlotRequestParams', {amp: true}, env => {
-  let doc, win;
-  beforeEach(() => {
-    win = env.win;
-    doc = win.document;
-  });
-
-  function createResource(
-    config,
-    layout,
-    tagName = 'amp-ad',
-    parent = doc.body
-  ) {
-    config['layout'] = layout;
-    const element = createElementWithAttributes(doc, tagName, config);
-    parent.appendChild(element);
-    return element;
-  }
-
-  it('should return the fixed width for FIXED layout', () => {
-    const element = createResource({width: 300, height: 250}, 'fixed');
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it('should return 0 for FIXED layout and invalid width', () => {
-    allowConsoleError(() => {
-      const element = createResource({width: 'auto', height: 250}, 'fixed');
-      expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
-    });
-  });
-
-  it('should return 0 for NODISPLAY layout', () => {
-    const element = createResource({width: 500}, 'nodisplay');
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
-  });
-
-  it('should return 0 for FLEX_ITEM layout', () => {
-    const element = createResource({width: 500}, 'flex-item');
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
-  });
-
-  it('should return 0 for invalid layout', () => {
-    allowConsoleError(() => {
-      const element = createResource({width: 500}, 'qwerty');
-      expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
-    });
-  });
-
-  it('should return the max-width, if present, for FILL layout', () => {
-    const element = createResource({maxWidth: 300}, 'fill');
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it("should return parent's fixed width for FILL layout", () => {
-    const parent = document.createElement('div');
-    parent.setAttribute('width', 300);
-    parent.setAttribute('layout', 'fixed');
-    doc.body.appendChild(parent);
-    const element = createResource({} /* config */, 'fill', 'amp-ad', parent);
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it('should return the max-width, if present, for FIXED_HEIGHT layout', () => {
-    const element = createResource({height: 300}, 'fixed-height');
-    element.style.maxWidth = '250px';
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(250);
-  });
-
-  it("should return parent's fixed width for FIXED_HEIGHT layout", () => {
-    const parent = document.createElement('div');
-    parent.setAttribute('width', 300);
-    parent.setAttribute('layout', 'fixed');
-    doc.body.appendChild(parent);
-    const element = createResource(
-      {height: 250},
-      'fixed-height',
-      'amp-ad',
-      parent
-    );
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it('should return the max-width, if present, for FLUID layout', () => {
-    const element = createResource({height: 300}, 'fluid');
-    element.style.maxWidth = '250px';
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(250);
-  });
-
-  it("should return parent's fixed width for FLUID layout", () => {
-    const parent = document.createElement('div');
-    parent.setAttribute('width', 300);
-    parent.setAttribute('layout', 'fixed');
-    doc.body.appendChild(parent);
-    const element = createResource({height: 250}, 'fluid', 'amp-ad', parent);
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it('should return the max-width, if present, for RESPONSIVE layout', () => {
-    const element = createResource({height: 200, width: 200}, 'responsive');
-    element.style.maxWidth = '250px';
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(250);
-  });
-
-  it("should return parent's fixed width for RESPONSIVE layout", () => {
-    const parent = document.createElement('div');
-    parent.setAttribute('width', 300);
-    parent.setAttribute('layout', 'fixed');
-    doc.body.appendChild(parent);
-    const element = createResource(
-      {height: 250, width: 250},
-      'responsive',
-      'amp-ad',
-      parent
-    );
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it('should return the viewport width for CONTAINER layout', () => {
-    const element = createResource({} /* config */, 'container');
-    sandbox
-      .stub(Services.viewportForDoc(element), 'getSize')
-      .returns({width: 300});
-    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
-  });
-
-  it('should return -1 width for non-fixed layouts, maxDepth = 1', () => {
-    ['fill', 'fixed-height', 'fluid', 'responsive'].forEach(layout => {
-      const parent = document.createElement('div');
-      parent.setAttribute('width', 300);
-      parent.setAttribute('layout', 'fixed');
-      doc.body.appendChild(parent);
-      const element = createResource({height: 250}, layout, 'amp-ad', parent);
-      expect(getFlexibleAdSlotRequestParams(win, element).msz).to.equal(-1);
     });
   });
 });
