+++ extensions/amp-ad-network-doubleclick-impl/0.1/test/test-flexible-ad-slot-utils.js
@@ -1,0 +1,156 @@
+/**
+ * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import {Services} from '../../../../src/services';
+import {createElementWithAttributes} from '../../../../src/dom';
+import {getFlexibleAdSlotRequestParams} from '../flexible-ad-slot-utils';
+
+
+describes.realWin('#getFlexibleAdSlotRequestParams', {amp: true}, env => {
+  let doc, win;
+  beforeEach(() => {
+    win = env.win;
+    doc = win.document;
+  });
+
+  function createResource(
+    config,
+    layout,
+    tagName = 'amp-ad',
+    parent = doc.body
+  ) {
+    config['layout'] = layout;
+    const element = createElementWithAttributes(doc, tagName, config);
+    parent.appendChild(element);
+    return element;
+  }
+
+  it('should return the fixed width for FIXED layout', () => {
+    const element = createResource({width: 300, height: 250}, 'fixed');
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it('should return 0 for FIXED layout and invalid width', () => {
+    allowConsoleError(() => {
+      const element = createResource({width: 'auto', height: 250}, 'fixed');
+      expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
+    });
+  });
+
+  it('should return 0 for NODISPLAY layout', () => {
+    const element = createResource({width: 500}, 'nodisplay');
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
+  });
+
+  it('should return 0 for FLEX_ITEM layout', () => {
+    const element = createResource({width: 500}, 'flex-item');
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
+  });
+
+  it('should return 0 for invalid layout', () => {
+    allowConsoleError(() => {
+      const element = createResource({width: 500}, 'qwerty');
+      expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(0);
+    });
+  });
+
+  it('should return the max-width, if present, for FILL layout', () => {
+    const element = createResource({maxWidth: 300}, 'fill');
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it("should return parent's fixed width for FILL layout", () => {
+    const parent = document.createElement('div');
+    parent.setAttribute('width', 300);
+    parent.setAttribute('layout', 'fixed');
+    doc.body.appendChild(parent);
+    const element = createResource({} /* config */, 'fill', 'amp-ad', parent);
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it('should return the max-width, if present, for FIXED_HEIGHT layout', () => {
+    const element = createResource({height: 300}, 'fixed-height');
+    element.style.maxWidth = '250px';
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(250);
+  });
+
+  it("should return parent's fixed width for FIXED_HEIGHT layout", () => {
+    const parent = document.createElement('div');
+    parent.setAttribute('width', 300);
+    parent.setAttribute('layout', 'fixed');
+    doc.body.appendChild(parent);
+    const element = createResource(
+      {height: 250},
+      'fixed-height',
+      'amp-ad',
+      parent
+    );
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it('should return the max-width, if present, for FLUID layout', () => {
+    const element = createResource({height: 300}, 'fluid');
+    element.style.maxWidth = '250px';
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(250);
+  });
+
+  it("should return parent's fixed width for FLUID layout", () => {
+    const parent = document.createElement('div');
+    parent.setAttribute('width', 300);
+    parent.setAttribute('layout', 'fixed');
+    doc.body.appendChild(parent);
+    const element = createResource({height: 250}, 'fluid', 'amp-ad', parent);
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it('should return the max-width, if present, for RESPONSIVE layout', () => {
+    const element = createResource({height: 200, width: 200}, 'responsive');
+    element.style.maxWidth = '250px';
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(250);
+  });
+
+  it("should return parent's fixed width for RESPONSIVE layout", () => {
+    const parent = document.createElement('div');
+    parent.setAttribute('width', 300);
+    parent.setAttribute('layout', 'fixed');
+    doc.body.appendChild(parent);
+    const element = createResource(
+      {height: 250, width: 250},
+      'responsive',
+      'amp-ad',
+      parent
+    );
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it('should return the viewport width for CONTAINER layout', () => {
+    const element = createResource({} /* config */, 'container');
+    sandbox
+      .stub(Services.viewportForDoc(element), 'getSize')
+      .returns({width: 300});
+    expect(getFlexibleAdSlotRequestParams(win, element).psz).to.equal(300);
+  });
+
+  it('should return null for non-fixed layouts', () => {
+    ['fill', 'fixed-height', 'fluid', 'responsive'].forEach(layout => {
+      const parent = document.createElement('div');
+      parent.setAttribute('width', 300);
+      parent.setAttribute('layout', 'fixed');
+      doc.body.appendChild(parent);
+      const element = createResource({height: 250}, layout, 'amp-ad', parent);
+      expect(getFlexibleAdSlotRequestParams(win, element).msz).to.be.null;
+    });
+  });
+});
