+++ test/unit/utils/test-dom-ancestor-visitor.js
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import {createElementWithAttributes} from '../../../../src/dom';
+import {getFlexibleAdSlotRequestParams} from '../flexible-ad-slot-utils';
+
+
+describes.realWin('#getFlexibleAdSlotRequestParams', {amp: true}, env => {
+  let doc, win;
+  beforeEach(() => {
+    win = env.win;
+    doc = win.document;
+  });
+
+  function createResource(
+    config,
+    layout,
+    tagName = 'amp-ad',
+    parent = doc.body
+  ) {
+    config['layout'] = layout;
+    const element = createElementWithAttributes(doc, tagName, config);
+    parent.appendChild(element);
+    return element;
+  }
+});
+
